+++ pcs/cli/common/lib_wrapper.py
@@ -46,7 +46,7 @@ def wrapper(dictionary):
 
 def cli_env_to_lib_env(cli_env):
     return LibraryEnvironment(
-        logging.getLogger("old_cli"),
+        logging.getLogger("pcs"),
         LibraryReportProcessorToConsole(cli_env.debug),
         cli_env.user,
         cli_env.groups,